Initial changes for 8254234 - Add test library stream object builder
